+++ src/library/fileio/stdiowrappers.cpp
@@ -38,6 +38,10 @@ DEFINE_ORIG_POINTER(fopen)
 DEFINE_ORIG_POINTER(fopen64)
 DEFINE_ORIG_POINTER(fclose)
 DEFINE_ORIG_POINTER(fileno)
+DEFINE_ORIG_POINTER(setbuf)
+DEFINE_ORIG_POINTER(setvbuf)
+DEFINE_ORIG_POINTER(setbuffer)
+DEFINE_ORIG_POINTER(setlinebuf)
 
 FILE *fopen (const char *filename, const char *modes)
 {
@@ -222,6 +226,74 @@ int fileno (FILE *stream) __THROW
     }
 
     return orig::fileno(stream);
+}
+
+void setbuf (FILE *stream, char *buf) __THROW
+{
+    LINK_NAMESPACE_GLOBAL(setbuf);
+
+    if (GlobalState::isNative())
+        return orig::setbuf(stream, buf);
+
+    LOGTRACE_SIMPLE(LCF_FILEIO);
+
+    /* If the stream is a savefile, keep the stream unbuffered. */
+    const SaveFile* sf = SaveFileList::getSaveFile(stream);
+    if (sf)
+        return;
+
+    return orig::setbuf(stream, buf);
+}
+
+int setvbuf (FILE *stream, char *buf, int modes, size_t n) __THROW
+{
+    LINK_NAMESPACE_GLOBAL(setvbuf);
+
+    if (GlobalState::isNative())
+        return orig::setvbuf(stream, buf, modes, n);
+
+    LOGTRACE(LCF_FILEIO, "%s call with stream %p and mode %d", __func__, stream, modes);
+
+    /* If the stream is a savefile, keep the stream unbuffered. */
+    const SaveFile* sf = SaveFileList::getSaveFile(stream);
+    if (sf)
+        return 0;
+
+    return orig::setvbuf(stream, buf, modes, n);
+}
+
+void setbuffer (FILE *stream, char *buf, size_t size) __THROW
+{
+    LINK_NAMESPACE_GLOBAL(setbuffer);
+
+    if (GlobalState::isNative())
+        return orig::setbuffer(stream, buf, size);
+
+    LOGTRACE_SIMPLE(LCF_FILEIO);
+
+    /* If the stream is a savefile, keep the stream unbuffered. */
+    const SaveFile* sf = SaveFileList::getSaveFile(stream);
+    if (sf)
+        return;
+
+    return orig::setbuffer(stream, buf, size);
+}
+
+void setlinebuf (FILE *stream) __THROW
+{
+    LINK_NAMESPACE_GLOBAL(setlinebuf);
+
+    if (GlobalState::isNative())
+        return orig::setlinebuf(stream);
+
+    LOGTRACE_SIMPLE(LCF_FILEIO);
+
+    /* If the stream is a savefile, keep the stream unbuffered. */
+    const SaveFile* sf = SaveFileList::getSaveFile(stream);
+    if (sf)
+        return;
+
+    return orig::setlinebuf(stream);
 }
 
 }